+++ CHANGES.md
@@ -213,6 +213,7 @@
 - Layout adjustment
 
 ### Pre-release (work-in-progress)
+- Phonak: Roger Upgrader update checker (in Extra Menu)
 - Extra Menu: text adjustment
 - Extra Menu: ReBrand Code Viewer
 - Pre-Release Checker: change filename with latest commit date and ID